Menyelesaikan bug di Text Input Email dan password pada activity_register.xml dan actvity_login.xml dan mengubha slogan
