--- app/src/main/java/com/example/capstoneproject/View/CustomView/MyEditText.kt
@@ -1,54 +1,0 @@
-package com.example.capstoneproject.View.CustomView
-
-import android.content.Context
-import android.graphics.Canvas
-import android.graphics.drawable.Drawable
-import android.text.Editable
-import android.text.TextWatcher
-import android.util.AttributeSet
-import android.util.Patterns
-import android.view.MotionEvent
-import android.view.View
-import androidx.appcompat.widget.AppCompatEditText
-import com.example.capstoneproject.R
-
-class MyEditText @JvmOverloads constructor (
-    context: Context,
-    attrs: AttributeSet? = null
-): AppCompatEditText(context,attrs), View.OnTouchListener {
-    init {
-        setOnTouchListener(this)
-        addTextChangedListener(object : TextWatcher {
-
-            override fun beforeTextChanged(s: CharSequence, start: Int, count: Int, after: Int) {}
-
-            override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {
-
-            }
-
-            override fun afterTextChanged(s: Editable?) {}
-        })
-    }
-
-    override fun onDraw(canvas: Canvas) {
-        super.onDraw(canvas)
-        textAlignment = View.TEXT_ALIGNMENT_VIEW_START
-    }
-
-    private fun setButtonDrawables(
-        startOfTheText: Drawable? = null,
-        topOfTheText: Drawable? = null,
-        endOfTheText: Drawable? = null,
-        bottomOfTheText: Drawable? = null
-    ){
-        setCompoundDrawablesWithIntrinsicBounds(
-            startOfTheText,
-            topOfTheText,
-            endOfTheText,
-            bottomOfTheText)
-    }
-
-    override fun onTouch(v: View?, event: MotionEvent?): Boolean {
-        return false
-    }
-}
