+++ app/src/main/java/com/example/capstoneproject/View/CustomView/MyPasswordCustom.kt
@@ -1,0 +1,4 @@
+package com.example.capstoneproject.View.CustomView
+
+class MyPasswordCustom {
+}
